@@ -8,12 +8,15 @@ const nextConfig = removeImports({
   images: {
     domains: [
       'lh3.googleusercontent.com',
-      'api.dicebear.com',
       'storage.googleapis.com',
     ],
   },
-  experimental: {
-    swcPlugins: [['@swc-jotai/react-refresh', {}]],
+  typescript: {
+    // !! WARN !!
+    // Dangerously allow production builds to successfully complete even if
+    // your project has type errors.
+    // !! WARN !!
+    ignoreBuildErrors: true,
   },
   webpack: (config) => {
     config.experiments = { ...config.experiments, topLevelAwait: true };
@@ -21,21 +24,21 @@ const nextConfig = removeImports({
   },
 });
 
-const sentryConfig = withSentryConfig(
-  nextConfig,
-  {
-    authToken: process.env.SENTRY_AUTH_TOKEN,
-    silent: true,
-    org: 'thibaud-brault',
-    project: 'faqmaker',
-  },
-  {
-    widenClientFileUpload: true,
-    transpileClientSDK: true,
-    tunnelRoute: '/monitoring',
-    hideSourceMaps: true,
-    disableLogger: true,
-  },
-);
+// const sentryConfig = withSentryConfig(
+//   nextConfig,
+//   {
+//     authToken: process.env.SENTRY_AUTH_TOKEN,
+//     silent: true,
+//     org: 'thibaud-brault',
+//     project: 'faqmaker',
+//   },
+//   {
+//     widenClientFileUpload: true,
+//     transpileClientSDK: true,
+//     tunnelRoute: '/monitoring',
+//     hideSourceMaps: true,
+//     disableLogger: true,
+//   },
+// );
 
-module.exports = sentryConfig;
+module.exports = nextConfig;
