@@ -1,23 +1,17 @@
+'use client';
+
 import { MoveLeft } from 'lucide-react';
-import { useRouter } from 'next/router';
+import Link from 'next/link';
 
 import { Button } from './Button';
 
 export const BackButton = () => {
-  const router = useRouter();
   return (
-    <Button
-      variant="primary"
-      weight="semibold"
-      icon="withIcon"
-      font="large"
-      size="small"
-      className="lowercase"
-      style={{ fontVariant: 'small-caps' }}
-      onClick={() => router.back()}
-    >
-      <MoveLeft />
-      Go back
+    <Button variant="primary" icon={true} font="large" size="small" asChild>
+      <Link href="/">
+        <MoveLeft />
+        Go back
+      </Link>
     </Button>
   );
 };
