@@ -1,67 +1,64 @@
-import { ButtonHTMLAttributes, forwardRef } from 'react';
+import type { ButtonHTMLAttributes } from 'react';
+import { forwardRef } from 'react';
 
 import { Slot } from '@radix-ui/react-slot';
 import { cva, type VariantProps } from 'class-variance-authority';
 
 import { cn } from '@/utils';
 
-const button = cva('transition-all duration-300', {
-  variants: {
-    variant: {
-      primary: [
-        'bg-teal-9 hover:bg-teal-10 text-white shadow-sm shadow-transparent',
-      ],
-      ghost: [
-        'bg-gray-3 hover:bg-gray-4 shadow-sm shadow-grayA-8 text-gray-12',
-      ],
-      secondary: [
-        'bg-transparent text-tealA-11 shadow shadow-sm shadow-tealA-7 hover:shadow-tealA-8',
-      ],
-      disabled: [
-        'bg-teal-surfaceLight dark:bg-teal-surfaceDark shadow-sm text-tealA-11 shadow-tealA-7 hover:shadow-tealA-8',
-      ],
-      disabledDestructive: [
-        'bg-red-surfaceLight dark:bg-red-surfaceDark shadow-sm text-redA-11 shadow-redA-7 hover:shadow-redA-8',
-      ],
-      destructive: [
-        'bg-red-9 hover:bg-red-10 text-white shadow-sm shadow-transparent',
-      ],
+export const buttonVariants = cva(
+  'transition-all duration-300 ease-in-out lowercase',
+  {
+    variants: {
+      variant: {
+        primary:
+          'bg-primary-negative text-primary-negative shadow-sm shadow-transparent disabled:bg-primary-negative-disabled disabled:text-primary disabled:shadow-sm disabled:shadow-primary disabled:hover:shadow-primary-hover',
+        ghost:
+          'bg-primary-foreground text-primary shadow-sm shadow-primary hover:bg-primary-foreground-hover disabled:shadow-transparent disabled:bg-transparent',
+
+        secondary:
+          'bg-transparent text-accent-secondary shadow-sm shadow-accent hover:shadow-accent-hover',
+
+        destructive:
+          'bg-destructive text-white shadow-sm shadow-transparent hover:bg-destructive-hover disabled:bg-destructive-disabled disabled:text-redA-11 disabled:shadow-sm disabled:shadow-destructive disabled:hover:shadow-destructive-hover disabled:dark:bg-destructive-disabled-dark',
+      },
+      icon: {
+        true: 'flex items-center justify-center gap-1',
+      },
+      font: {
+        small: 'text-sm',
+        base: 'text-base',
+        large: 'text-lg',
+      },
+      rounded: {
+        none: 'rounded-none',
+        base: 'rounded-md',
+        bottom: 'rounded-b-md',
+        full: 'rounded-full',
+      },
+      weight: {
+        bold: 'font-bold',
+        semibold: 'font-semibold',
+      },
+      size: {
+        small: 'w-fit px-2 py-1',
+        medium: 'w-fit px-4 py-2',
+        full: 'w-full py-2',
+        icon: 'size-6',
+      },
     },
-    icon: {
-      withIcon: ['flex justify-center items-center gap-1'],
-    },
-    font: {
-      small: ['text-sm'],
-      base: ['text-base'],
-      large: ['text-lg'],
-    },
-    rounded: {
-      none: ['rounded-none'],
-      base: ['rounded-md'],
-      bottom: ['rounded-b-md'],
-      full: ['rounded-full'],
-    },
-    weight: {
-      bold: ['font-bold'],
-      semibold: ['font-semibold'],
-    },
-    size: {
-      small: ['py-1 px-2 w-fit'],
-      medium: ['py-2 px-4 w-fit'],
-      full: ['py-2 w-full'],
-      icon: ['h-6 w-6'],
+    defaultVariants: {
+      font: 'base',
+      rounded: 'base',
+      size: 'medium',
+      weight: 'semibold',
     },
   },
-  defaultVariants: {
-    font: 'base',
-    rounded: 'base',
-    size: 'medium',
-  },
-});
+);
 
 export interface ButtonProps
   extends ButtonHTMLAttributes<HTMLButtonElement>,
-    VariantProps<typeof button> {
+    VariantProps<typeof buttonVariants> {
   asChild?: boolean;
 }
 
@@ -83,8 +80,9 @@ export const Button = forwardRef<HTMLButtonElement, ButtonProps>(
     const Comp = asChild ? Slot : 'button';
     return (
       <Comp
+        style={{ fontVariant: 'small-caps' }}
         className={cn(
-          button({
+          buttonVariants({
             variant,
             font,
             icon,
