@@ -1,3 +1,5 @@
+'use client';
+
 import '@uiw/react-md-editor/markdown-editor.css';
 
 import dynamic from 'next/dynamic';
@@ -8,6 +10,7 @@ import { useMediaQuery } from '@/hooks';
 const MDEditor = dynamic(() => import('@uiw/react-md-editor'), { ssr: false });
 
 type Props = {
+  /* eslint-disable-next-line @typescript-eslint/no-explicit-any*/
   onChange: (...event: any[]) => void;
   value: string;
 };
@@ -18,7 +21,7 @@ export const Editor = ({ onChange, value }: Props) => {
   return (
     <div className="w-full">
       <MDEditor
-        className="!w-full !rounded-md !bg-gray-3 !text-gray-12 !shadow-sm [&_.w-md-editor-toolbar]:!rounded-t-md [&_.w-md-editor-toolbar]:!bg-gray-2"
+        className="!w-full !rounded-md !bg-primary-foreground !text-primary !shadow-sm [&_.w-md-editor-toolbar]:!rounded-t-md [&_.w-md-editor-toolbar]:!bg-transparent"
         value={value}
         preview={isDesktop ? 'live' : 'edit'}
         onChange={onChange}
