@@ -1,3 +1,4 @@
+export * from './alert';
 export * from './avatar';
 export * from './badge';
 export * from './button';
@@ -5,16 +6,15 @@ export * from './dialog';
 export * from './drawer';
 export * from './dropdown';
 export * from './editor';
-export * from './email';
-export * from './error';
+export * from './emails';
 export * from './field';
 export * from './input';
 export * from './label';
 export * from './loader';
 export * from './pagination';
+export * from './popover';
 export * from './select';
 export * from './stepper';
 export * from './tabs';
-export * from './theme';
 export * from './toast';
 export * from './tooltip';
