@@ -1,5 +1,3 @@
-import React from 'react';
-
 type Props = {
   company: string;
 };
@@ -8,7 +6,7 @@ export const Footer = ({ company }: Props) => {
   const year = new Date().getFullYear();
 
   return (
-    <footer className="flex w-full items-center justify-between border-t border-t-gray-6 bg-gray-2 px-8 py-2 text-xs text-gray-12">
+    <footer className="flex w-full items-center justify-between border-t border-t-gray-6 bg-primary-foreground px-8 py-2 text-xs text-primary">
       <p>© {year} FAQMaker</p>
       <p className="text-sm font-bold">{company}</p>
       <a href="mailto:contact@faqmaker.co" className="hover:underline">
