@@ -1,6 +1,7 @@
 export * from './footer';
 export * from './header';
-export * from './list';
+export * from './home';
+export * from './navigation';
 export * from './profile';
 export * from './question';
 export * from './search';
