@@ -1,54 +1,55 @@
-import { MoveRight } from 'lucide-react';
+'use client';
+
 import dynamic from 'next/dynamic';
 import Link from 'next/link';
 
-import { Loader } from '@/components';
-import { NodeWithQuestionAndAnswer } from '@/hooks';
+import { dateOptions, Routes } from '@/utils';
+
+import type { NodeWithQuestionAndAnswer } from '@/types';
+
 const MarkdownPreview = dynamic(() => import('@uiw/react-markdown-preview'), {
   ssr: false,
 });
 
 type Props = {
   nodes?: NodeWithQuestionAndAnswer[];
-  isPending: boolean;
 };
 
-export const UserAnswers = ({ nodes, isPending }: Props) => {
-  if (isPending) {
-    return <Loader size="items" />;
-  }
-
+export const UserAnswers = ({ nodes }: Props) => {
+  console.log(nodes);
   return (
-    <>
+    <section className="space-y-4">
       <h2
-        className="text-center font-serif text-3xl font-semibold lowercase md:text-4xl"
+        className="text-xl font-semibold lowercase"
         style={{ fontVariant: 'small-caps' }}
       >
         Answers
       </h2>
       {nodes && nodes.length > 0 ? (
-        <ul className="flex list-none flex-col gap-2 ">
-          {nodes.map((node, index) => (
-            <li
-              className="flex items-center justify-between rounded-md px-3 py-2 shadow-sm"
-              key={index}
+        <div className="flex flex-col gap-2">
+          {nodes.map((node) => (
+            <Link
+              href={`${Routes.SITE.QUESTION.INDEX}/${node.question.id}`}
+              className="flex items-center justify-between rounded-md px-3 py-2 hover:bg-primary-foreground-hover"
+              key={node.id}
             >
-              <MarkdownPreview source={node.answer.text} />
-
-              <Link
-                href={{
-                  pathname: '/question/[id]',
-                  query: { id: node.question.id },
-                }}
-              >
-                <MoveRight />
-              </Link>
-            </li>
+              <MarkdownPreview
+                source={node.answer.text}
+                className="!bg-transparent"
+              />
+              <small className="text-primary-muted">
+                Answered on{' '}
+                {new Date(node.answer.updatedAt).toLocaleDateString(
+                  undefined,
+                  dateOptions,
+                )}
+              </small>
+            </Link>
           ))}
-        </ul>
+        </div>
       ) : (
         <p className="text-center italic">No answers</p>
       )}
-    </>
+    </section>
   );
 };
