@@ -1,63 +1,46 @@
-import { Question } from '@prisma/client';
+'use client';
+
 import Link from 'next/link';
 
-import { Loader } from '@/components';
-import { dateOptions } from '@/utils';
+import { Routes, dateOptions } from '@/utils';
 
-type QuestionWithNodeId = Question & {
-  node: {
-    id: string;
-  };
-};
+import type { QuestionWithNodeId } from '@/types';
 
 type Props = {
   questions?: QuestionWithNodeId[];
-  isPending: boolean;
 };
 
-export const UserQuestions = ({ questions, isPending }: Props) => {
-  if (isPending) {
-    return <Loader size="items" />;
-  }
-
+export const UserQuestions = ({ questions }: Props) => {
   return (
-    <>
+    <section className="space-y-4">
       <h2
-        className="text-center font-serif text-3xl font-semibold lowercase md:text-4xl"
+        className="text-xl font-semibold lowercase"
         style={{ fontVariant: 'small-caps' }}
       >
         Questions
       </h2>
       {questions && questions.length > 0 ? (
-        <ul className="flex list-none flex-col gap-2">
+        <div className="flex flex-col gap-2">
           {questions?.map((question) => (
-            <li
-              className="flex items-center justify-between rounded-md px-3 py-2 shadow-sm"
+            <Link
+              className="flex items-center justify-between rounded-md px-3 py-2 hover:bg-primary-foreground-hover"
               key={question.id}
+              href={`${Routes.SITE.QUESTION.INDEX}/${question.node.id}`}
             >
-              <h3 className="text-xl font-semibold hover:underline md:text-2xl">
-                <Link
-                  href={{
-                    pathname: '/question/[id]',
-                    query: { id: question.node.id },
-                  }}
-                >
-                  {question.text}
-                </Link>
-              </h3>
-              <p className="text-xs">
+              <h3 className="font-semibold">{question.text}</h3>
+              <small className="text-primary-muted">
                 Asked on{' '}
                 {new Date(question.createdAt).toLocaleDateString(
                   undefined,
                   dateOptions,
                 )}
-              </p>
-            </li>
+              </small>
+            </Link>
           ))}
-        </ul>
+        </div>
       ) : (
         <p className="text-center italic">No questions</p>
       )}
-    </>
+    </section>
   );
 };
