@@ -1,21 +1,28 @@
+'use client';
+
 import { useEffect, useMemo, useState } from 'react';
 
 import { zodResolver } from '@hookform/resolvers/zod';
 import { AtSign, UserIcon } from 'lucide-react';
 import Image from 'next/image';
-import { SubmitHandler, useForm } from 'react-hook-form';
-import { z } from 'zod';
+import { useForm } from 'react-hook-form';
 
-import { Button, Field, Input } from '@/components';
-import { useMediaQuery, useUpdateUser } from '@/hooks';
-import { updateUserClientSchema } from '@/lib';
-import { IUserUpdateFields, UserWithTenant } from '@/types';
+import { updateUser, updateUserSchema } from '@/actions';
+import { Button, Field, Input, resultToast } from '@/components';
+import { useMediaQuery } from '@/hooks';
+import { Limits } from '@/utils';
+
+import { PageChangeAlert } from '../navigation';
+
+import type { IUserUpdateFields, Me } from '@/types';
+import type { SubmitHandler } from 'react-hook-form';
+import type { z } from 'zod';
 
 type Props = {
-  me: UserWithTenant;
+  me: Me;
 };
 
-type Schema = z.infer<typeof updateUserClientSchema>;
+type Schema = z.infer<typeof updateUserSchema>;
 
 export const UpdateProfile = ({ me }: Props) => {
   const [disabled, setDisabled] = useState<boolean>(true);
@@ -24,20 +31,23 @@ export const UpdateProfile = ({ me }: Props) => {
   const {
     register,
     handleSubmit,
+    watch,
     formState: { isSubmitting, isDirty, errors, isValid },
   } = useForm<Schema>({
-    resolver: zodResolver(updateUserClientSchema),
+    resolver: zodResolver(updateUserSchema),
     mode: 'onBlur',
     defaultValues: {
       name: me.name ?? '',
       email: me.email,
+      role: me.role,
+      tenantId: me.tenantId,
+      id: me.id,
     },
   });
 
-  const { mutate } = useUpdateUser(me.id, me.tenantId);
-
-  const onSubmit: SubmitHandler<Schema> = (values) => {
-    mutate(values);
+  const onSubmit: SubmitHandler<Schema> = async (data) => {
+    const result = await updateUser(data);
+    resultToast(result?.serverError, result?.data?.message);
   };
 
   const fields: IUserUpdateFields[] = useMemo(
@@ -46,13 +56,15 @@ export const UpdateProfile = ({ me }: Props) => {
         label: 'Name',
         value: 'name',
         type: 'text',
-        icon: <UserIcon className="h-5 w-5" />,
+        limit: Limits.NAME,
+        icon: <UserIcon className="size-5" />,
       },
       {
         label: 'Email',
         value: 'email',
         type: 'email',
-        icon: <AtSign className="h-5 w-5" />,
+        limit: Limits.EMAIL,
+        icon: <AtSign className="size-5" />,
       },
     ],
     [],
@@ -63,67 +75,64 @@ export const UpdateProfile = ({ me }: Props) => {
   }, [isDirty, isSubmitting, isValid]);
 
   return (
-    <form
-      onSubmit={handleSubmit(onSubmit)}
-      className="flex flex-col items-center gap-4"
-    >
-      <fieldset className="flex w-full flex-col gap-4 sm:grid sm:grid-cols-4">
-        <div className="col-span-4 w-full text-center">
-          <legend
-            className="text-center font-serif text-3xl font-semibold lowercase md:text-4xl"
-            style={{ fontVariant: 'small-caps' }}
-          >
-            Informations
-          </legend>
-        </div>
-        {me.image ? (
-          <Image
-            src={me.image}
-            alt={'Profile picture'}
-            width={isDesktop ? 128 : 80}
-            height={isDesktop ? 128 : 80}
-            className="row-start-2 self-center justify-self-center rounded-md"
-          />
-        ) : (
-          <UserIcon className="row-start-2 h-20 w-20 self-center justify-self-center rounded-md sm:h-32 sm:w-32" />
-        )}
-        <ul className="col-span-3 row-start-2 flex list-none flex-col gap-2">
-          {fields.map((field) => (
-            <li key={field.value} className="flex flex-col gap-1">
-              <Field
-                label={field.label}
-                value={field.value}
-                error={errors[field.value]?.message}
-              >
-                <Input
-                  {...register(field.value)}
-                  defaultValue={me[field.value] ?? ''}
-                  withIcon
-                  icon={field.icon}
-                  type={field.type}
-                  id={field.value}
-                  placeholder={field.label}
-                />
-              </Field>
-            </li>
-          ))}
-          <li>
-            <span className="lowercase" style={{ fontVariant: 'small-caps' }}>
-              Role:
-            </span>{' '}
-            <b className="capitalize">{me.role}</b>
-          </li>
-        </ul>
-      </fieldset>
-      <Button
-        variant={disabled ? 'disabled' : 'primary'}
-        weight="semibold"
-        className="lowercase"
+    <section className="space-y-4">
+      <h2
+        className="text-xl font-semibold lowercase"
         style={{ fontVariant: 'small-caps' }}
-        disabled={disabled}
       >
-        Update
-      </Button>
-    </form>
+        Profile
+      </h2>
+      <form
+        onSubmit={handleSubmit(onSubmit)}
+        className="flex flex-col items-center gap-4"
+      >
+        <fieldset className="flex w-full gap-8">
+          {me.image ? (
+            <Image
+              src={me.image}
+              alt="Profile picture"
+              width={isDesktop ? 128 : 80}
+              height={isDesktop ? 128 : 80}
+              className="row-start-2 self-center justify-self-center rounded-md"
+            />
+          ) : (
+            <UserIcon className="row-start-2 size-20 self-center justify-self-center rounded-md sm:size-32" />
+          )}
+          <ul className="flex grow list-none flex-col gap-2">
+            {fields.map((field) => (
+              <li key={field.value} className="flex flex-col gap-1">
+                <Field
+                  label={field.label}
+                  value={field.value}
+                  error={errors[field.value]?.message}
+                  curLength={watch(field.value)?.length}
+                  limit={field.limit}
+                >
+                  <Input
+                    {...register(field.value)}
+                    defaultValue={me[field.value] ?? ''}
+                    withIcon
+                    icon={field.icon}
+                    type={field.type}
+                    id={field.value}
+                    placeholder={field.label}
+                  />
+                </Field>
+              </li>
+            ))}
+            <li>
+              <span className="lowercase" style={{ fontVariant: 'small-caps' }}>
+                Role:
+              </span>{' '}
+              <b className="capitalize">{me.role}</b>
+            </li>
+          </ul>
+        </fieldset>
+        <Button variant="primary" disabled={disabled}>
+          Update
+        </Button>
+      </form>
+      <PageChangeAlert isDirty={isDirty} />
+    </section>
   );
 };
