@@ -1,84 +1,73 @@
-import { useEffect, useMemo, useState } from 'react';
+'use client';
+
+import { Fragment, useEffect, useState } from 'react';
 
 import { zodResolver } from '@hookform/resolvers/zod';
-import { Tenant } from '@prisma/client';
-import { useRouter } from 'next/router';
-import { SubmitHandler, useForm } from 'react-hook-form';
-import { z } from 'zod';
+import { useForm } from 'react-hook-form';
 
-import { Button, Field, Input } from '@/components';
-import { useUpdateTenant } from '@/hooks';
-import { updateTenantClientSchema } from '@/lib';
-import { ITenantUpdateFields } from '@/types';
+import { updateTenant, updateTenantSchema } from '@/actions';
+import { Button, Field, Input, resultToast } from '@/components';
+import { PageChangeAlert } from '@/modules';
+import { Limits } from '@/utils';
+
+import type { ITenantUpdateFields } from '@/types';
+import type { Integrations, Tenant } from '@prisma/client';
+import type { SubmitHandler } from 'react-hook-form';
+import type { z } from 'zod';
 
 type Props = {
   tenant: Tenant;
+  integrations: Integrations | null;
 };
 
-type Schema = z.infer<typeof updateTenantClientSchema>;
+type Schema = z.infer<typeof updateTenantSchema>;
 
-export const Company = ({ tenant }: Props) => {
+export function Company({ tenant, integrations }: Props) {
   const [disabled, setDisabled] = useState<boolean>(true);
-  const router = useRouter();
   const {
     register,
     handleSubmit,
     watch,
-    reset,
     formState: { isSubmitting, isDirty, isValid, errors },
   } = useForm<Schema>({
-    resolver: zodResolver(updateTenantClientSchema),
+    resolver: zodResolver(updateTenantSchema),
     mode: 'onBlur',
     defaultValues: {
       company: tenant.company,
       email: tenant.email,
-      domain: tenant.domain,
+      slack: integrations?.slack ?? '',
+      id: tenant.id,
     },
   });
-  const domainValue = watch('domain');
 
-  const { mutate } = useUpdateTenant(tenant.id, router);
-
-  const onSubmit: SubmitHandler<Schema> = (values) => {
-    mutate(values);
+  const onSubmit: SubmitHandler<Schema> = async (data) => {
+    const result = await updateTenant(data);
+    resultToast(result?.serverError, 'Tenant updated successfully');
   };
 
-  const fields: ITenantUpdateFields[] = useMemo(
-    () => [
-      {
-        label: 'Company',
-        value: 'company',
-        type: 'text',
-      },
-      {
-        label: 'Email',
-        value: 'email',
-        type: 'email',
-      },
-      {
-        label: 'Domain',
-        value: 'domain',
-        type: 'text',
-      },
-    ],
-    [],
-  );
+  const fields: ITenantUpdateFields[] = [
+    {
+      label: 'Company',
+      value: 'company',
+      type: 'text',
+      limit: Limits.COMPANY,
+    },
+    {
+      label: 'Email',
+      value: 'email',
+      type: 'email',
+      limit: Limits.EMAIL,
+    },
+  ];
 
   useEffect(() => {
-    if (domainValue === '') {
-      reset({
-        domain: tenant.domain,
-      });
-    }
     setDisabled(isSubmitting || !isDirty || !isValid);
-
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [isDirty, isSubmitting, isValid, domainValue]);
+  }, [isDirty, isSubmitting, isValid]);
 
   return (
-    <div className="flex flex-col gap-4">
+    <div className="space-y-4">
       <h2
-        className="text-center font-serif text-3xl font-semibold lowercase md:text-4xl"
+        className="text-xl font-semibold lowercase"
         style={{ fontVariant: 'small-caps' }}
       >
         Company
@@ -87,34 +76,43 @@ export const Company = ({ tenant }: Props) => {
         className="flex flex-col items-center gap-4"
         onSubmit={handleSubmit(onSubmit)}
       >
-        <fieldset className="flex w-full flex-col gap-2 md:flex-row">
-          {fields.map((field) => (
-            <div key={field.value} className="flex flex-1 flex-col">
-              <Field
-                label={field.label}
-                value={field.value}
-                error={errors[field.value]?.message}
-              >
+        <fieldset className="flex w-full flex-col">
+          <div className="space-y-1">
+            {fields.map((field) => (
+              <Fragment key={field.value}>
+                <Field
+                  label={field.label}
+                  value={field.value}
+                  error={errors[field.value]?.message}
+                  curLength={watch(field.value)?.length}
+                  limit={field.limit}
+                >
+                  <Input
+                    {...register(field.value)}
+                    type={field.type}
+                    id={field.value}
+                    placeholder={field.label}
+                  />
+                </Field>
+              </Fragment>
+            ))}
+            {tenant.plan !== 'free' && (
+              <Field label="Slack" value="slack" error={errors.slack?.message}>
                 <Input
-                  {...register(field.value)}
-                  type={field.type}
-                  id={field.value}
-                  placeholder={field.label}
+                  {...register('slack')}
+                  type="text"
+                  id="slack"
+                  placeholder="https://hooks.slack.com/services/"
                 />
               </Field>
-            </div>
-          ))}
+            )}
+          </div>
         </fieldset>
-        <Button
-          variant={disabled ? 'disabled' : 'primary'}
-          weight="semibold"
-          className="lowercase"
-          style={{ fontVariant: 'small-caps' }}
-          disabled={disabled}
-        >
+        <Button variant="primary" disabled={disabled}>
           Update
         </Button>
       </form>
+      <PageChangeAlert isDirty={isDirty} />
     </div>
   );
-};
+}
