@@ -1,45 +1,67 @@
+'use client';
+
 import { useEffect, useState } from 'react';
 
 import { zodResolver } from '@hookform/resolvers/zod';
-import { Tenant } from '@prisma/client';
 import { Upload } from 'lucide-react';
 import Image from 'next/image';
 import Dropzone from 'react-dropzone';
-import { Controller, SubmitHandler, useForm } from 'react-hook-form';
-import { z } from 'zod';
+import { Controller, useForm } from 'react-hook-form';
 
-import { Button } from '@/components';
-import { useUpsertFiles } from '@/hooks';
-import { filesClientSchema } from '@/lib';
+import { getSignedLogo, updateLogo, updateLogoSchema } from '@/actions';
+import { Button, resultToast } from '@/components';
+import { filesSchema } from '@/lib/validations';
+
+import type { Tenant } from '@prisma/client';
+import type { SubmitHandler } from 'react-hook-form';
+import type { z } from 'zod';
 
 type Props = {
   tenant: Tenant;
 };
 
-type Schema = z.infer<typeof filesClientSchema>;
+type Schema = z.infer<typeof filesSchema>;
+type UpdateLogoType = z.infer<typeof updateLogoSchema>;
 
 export const Files = ({ tenant }: Props) => {
   const [disabled, setDisabled] = useState<boolean>(true);
   const [previewImage, setPreviewImage] = useState<string>('');
-  const [file, setFile] = useState<File>();
+  const [file, setFile] = useState<File | null>(null);
 
   const {
     handleSubmit,
     control,
     reset,
-    formState: { isSubmitting, isDirty, isValid, errors },
+    formState: { isSubmitting, isDirty, isValid },
   } = useForm<Schema>({
-    resolver: zodResolver(filesClientSchema),
+    resolver: zodResolver(filesSchema),
     mode: 'onBlur',
     defaultValues: {
-      logo: null,
+      logo: undefined,
     },
   });
 
-  const { mutate } = useUpsertFiles(tenant.id);
-
-  const onSubmit: SubmitHandler<Schema> = (values) => {
-    mutate(values);
+  const onSubmit: SubmitHandler<Schema> = async (data) => {
+    const formData = new FormData();
+    const randomId = crypto.randomUUID();
+    const logo = encodeURIComponent(randomId + data.logo.name);
+    formData.append('logo', logo);
+    const { url, fields } = await getSignedLogo(formData);
+    formData.delete('logo');
+    Object.entries({ ...fields, file }).forEach(([key, value]) => {
+      formData.append(key, value as string | Blob);
+    });
+    await fetch(url, { method: 'POST', body: formData });
+    formData.forEach((_value, key) => {
+      formData.delete(key);
+    });
+    const logoUrl = `${url}logos/${logo}`;
+    const logoData: UpdateLogoType = {
+      logoUrl,
+      id: tenant.id,
+    };
+    const result = await updateLogo(logoData);
+    resultToast(result?.serverError, 'Logo updated successfully');
   };
 
   const handleReset = () => {
@@ -50,24 +72,22 @@ export const Files = ({ tenant }: Props) => {
 
   useEffect(() => {
     setDisabled(isSubmitting || !isDirty || !isValid);
-
-    // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [isDirty, isSubmitting, isValid]);
 
   return (
-    <div className="flex flex-col items-center gap-4">
+    <div className="flex flex-col gap-4">
       <h2
-        className="text-center font-serif text-3xl font-semibold lowercase md:text-4xl"
+        className="text-xl font-semibold lowercase"
         style={{ fontVariant: 'small-caps' }}
       >
         Logo
       </h2>
       <form
-        className="mx-auto flex w-11/12 flex-col items-center gap-4"
+        className="flex flex-col items-center gap-4"
         onSubmit={handleSubmit(onSubmit)}
       >
         <Controller
-          name={'logo'}
+          name="logo"
           control={control}
           render={({ field: { onChange, onBlur } }) => (
             <div className="w-full">
@@ -83,7 +103,7 @@ export const Files = ({ tenant }: Props) => {
               >
                 {({ getRootProps, getInputProps, open, isDragActive }) => (
                   <div
-                    className={`relative flex w-full cursor-pointer flex-col items-center justify-center gap-2 rounded-t-md border-2 border-b-0 border-dashed border-gray-12 px-4 py-8 ${isDragActive ? 'bg-grayA-3' : 'bg-transparent'}`}
+                    className={`relative flex w-full cursor-pointer flex-col items-center justify-center gap-2 rounded-t-md border-2 border-b-0 border-dashed border-gray-12 px-4 py-8 ${isDragActive ? 'bg-primary-foreground-alpha' : 'bg-transparent'}`}
                     {...getRootProps()}
                   >
                     <input
@@ -96,14 +116,14 @@ export const Files = ({ tenant }: Props) => {
                     {previewImage ? (
                       <Image
                         src={previewImage}
-                        alt={file.name}
-                        className="h-36 w-36 rounded-md border border-gray-7 object-cover"
+                        alt={file?.name ?? 'Logo'}
+                        className="size-36 rounded-md border border-primary object-cover"
                         width={144}
                         height={144}
                       />
                     ) : (
-                      <div className="flex h-36 w-36 items-center justify-center">
-                        <Upload className="h-20 w-20" />
+                      <div className="flex size-36 items-center justify-center">
+                        <Upload className="size-20" />
                       </div>
                     )}
                     <p className="text-xl font-semibold">
@@ -111,7 +131,7 @@ export const Files = ({ tenant }: Props) => {
                     </p>
                     <button
                       type="button"
-                      className="text-sm text-gray-11 hover:text-gray-12"
+                      className="text-sm text-primary-muted hover:text-primary"
                       onClick={open}
                     >
                       Choose a file or drag and drop
@@ -122,9 +142,8 @@ export const Files = ({ tenant }: Props) => {
               <Button
                 variant="ghost"
                 rounded="bottom"
-                weight="semibold"
-                className="w-full border-2 border-t border-dashed border-t-grayA-8 lowercase shadow-none"
-                style={{ fontVariant: 'small-caps', borderTopStyle: 'solid' }}
+                className="w-full border-2 border-t border-dashed border-t-grayA-8 shadow-none"
+                style={{ borderTopStyle: 'solid' }}
                 onClick={handleReset}
                 type="button"
               >
@@ -133,13 +152,7 @@ export const Files = ({ tenant }: Props) => {
             </div>
           )}
         />
-        <Button
-          variant={disabled ? 'disabled' : 'primary'}
-          weight="semibold"
-          className="lowercase"
-          style={{ fontVariant: 'small-caps' }}
-          disabled={disabled}
-        >
+        <Button variant="primary" disabled={disabled}>
           Update
         </Button>
       </form>
