@@ -1,11 +1,12 @@
+'use client';
+
 import { useEffect, useState } from 'react';
 
 import { zodResolver } from '@hookform/resolvers/zod';
 import { Flame } from 'lucide-react';
-import { useRouter } from 'next/router';
-import { SubmitHandler, useForm } from 'react-hook-form';
-import { z } from 'zod';
+import { useForm } from 'react-hook-form';
 
+import { deleteTenant, deleteTenantSchema } from '@/actions';
 import {
   Button,
   Dialog,
@@ -17,21 +18,18 @@ import {
   DialogTrigger,
   Input,
 } from '@/components';
-import { useDeleteTenant } from '@/hooks';
+
+import type { SubmitHandler } from 'react-hook-form';
+import type { z } from 'zod';
 
 type Props = {
   tenantId: string;
   company: string;
 };
+type Schema = z.infer<typeof deleteTenantSchema>;
 
 export const Delete = ({ tenantId, company }: Props) => {
-  const deleteSchema = z.object({
-    text: z.literal(`DELETE ${company}`),
-  });
-  type Schema = z.infer<typeof deleteSchema>;
-
   const [disabled, setDisabled] = useState<boolean>(true);
-  const router = useRouter();
 
   const {
     register,
@@ -39,15 +37,17 @@ export const Delete = ({ tenantId, company }: Props) => {
     setValue,
     formState: { isSubmitting, isValid },
   } = useForm<Schema>({
-    resolver: zodResolver(deleteSchema),
+    resolver: zodResolver(deleteTenantSchema),
     mode: 'onBlur',
+    defaultValues: {
+      text: '',
+      id: tenantId,
+      company,
+    },
   });
 
-  const { mutate } = useDeleteTenant(tenantId, company, router);
-
-  const onSubmit: SubmitHandler<Schema> = (values) => {
-    // @ts-ignore
-    mutate(values);
+  const onSubmit: SubmitHandler<Schema> = async (data) => {
+    await deleteTenant(data);
   };
 
   useEffect(() => {
@@ -57,13 +57,7 @@ export const Delete = ({ tenantId, company }: Props) => {
   return (
     <Dialog>
       <DialogTrigger asChild>
-        <Button
-          variant="destructive"
-          weight="semibold"
-          className="lowercase"
-          icon="withIcon"
-          style={{ fontVariant: 'small-caps' }}
-        >
+        <Button variant="destructive" icon={true}>
           <Flame />
           Delete account
         </Button>
@@ -77,7 +71,7 @@ export const Delete = ({ tenantId, company }: Props) => {
           account is permanent and will delete all your data forever.
         </p>
 
-        <p className="text-sm text-gray-11">
+        <p className="text-sm text-primary-muted">
           Type <span className="font-semibold">DELETE {company}</span> to
           confirm
         </p>
@@ -87,24 +81,13 @@ export const Delete = ({ tenantId, company }: Props) => {
             <DialogClose asChild>
               <Button
                 variant="ghost"
-                weight="semibold"
-                className="lowercase"
-                style={{ fontVariant: 'small-caps' }}
                 type="button"
-                // @ts-ignore
                 onClick={() => setValue('text', '')}
               >
                 Cancel
               </Button>
             </DialogClose>
-            <Button
-              variant={disabled ? 'disabledDestructive' : 'destructive'}
-              weight="semibold"
-              className="lowercase"
-              style={{ fontVariant: 'small-caps' }}
-              type="submit"
-              disabled={disabled}
-            >
+            <Button variant="destructive" type="submit" disabled={disabled}>
               Delete
             </Button>
           </DialogFooter>
