@@ -1,3 +1,5 @@
+'use client';
+
 import { SunIcon, MoonIcon } from 'lucide-react';
 import { useTheme } from 'next-themes';
 
@@ -30,11 +32,11 @@ export const ThemeToggle = () => {
   return (
     <DropdownMenu>
       <DropdownMenuTrigger asChild>
-        <Button size="icon" className="text-gray-12 hover:text-gray-11">
+        <Button size="icon" className="text-primary hover:text-primary-muted">
           {theme === 'light' ? (
-            <SunIcon className="h-6 w-6" />
+            <SunIcon className="size-6" />
           ) : (
-            <MoonIcon className="h-6 w-6" />
+            <MoonIcon className="size-6" />
           )}
           <span className="sr-only">Toggle theme</span>
         </Button>
@@ -45,8 +47,8 @@ export const ThemeToggle = () => {
             key={th.value}
             className={
               theme === th.value
-                ? 'bg-gray-12 text-white hover:bg-gray-12 dark:bg-gray-1 dark:hover:bg-gray-1'
-                : 'bg-gray-1 dark:bg-gray-12'
+                ? 'bg-primary-negative text-primary-negative hover:bg-primary-negative'
+                : 'bg-primary'
             }
             onClick={() => setTheme(th.value)}
           >
